mainpage1/requirements.txt
fastapi~=0.110.0
passlib~=1.7.4
pydantic~=2.6.3
SQLAlchemy~=2.0.27
starlette~=0.36.3
uvicorn~=0.27.1
python-dotenv~=1.0.1
openai~=1.12.0
python-jose==3.3.0
psycopg2==2.9.9
python-multipart==0.0.9
bcrypt==4.1.2